obsolete abeja.train.InstanceType
